fix: improve json parse error handling
